url-select: Unset underline flag after partial redraw
This fixes the behaviour where after redraw of only parts of the line
the underline flag was still being set even though the new line doesn't
match anymore. This can be simulated for example by starting an editor,
typing a URL and changing e.g. the http:// part to xttp:// without
changing the line length (vim: rx) where the part beginning with ttp://
keeps being underlined.

This commit sets the 3rd custom rendition bit (starting at 0) of a cell
iff the underline was being added as a result of the cell being part of
a match. If so, on a line update these underline flags get cleared again
(together with the custom rendition bit) to be then re-set after a
(possible) successful match.